@@ -15,41 +15,41 @@
 	},
 	"dependencies": {
 		"@formkit/auto-animate": "^0.8.2",
-		"@fortawesome/fontawesome-svg-core": "^6.6.0",
-		"@fortawesome/free-brands-svg-icons": "^6.6.0",
-		"@fortawesome/free-regular-svg-icons": "^6.6.0",
-		"@fortawesome/free-solid-svg-icons": "^6.6.0",
-		"@sveltejs/adapter-cloudflare": "^4.7.2",
-		"@unocss/preset-rem-to-px": "^0.63.4",
-		"@unocss/reset": "^0.63.3",
+		"@fortawesome/fontawesome-svg-core": "^6.7.1",
+		"@fortawesome/free-brands-svg-icons": "^6.7.1",
+		"@fortawesome/free-regular-svg-icons": "^6.7.1",
+		"@fortawesome/free-solid-svg-icons": "^6.7.1",
+		"@sveltejs/adapter-cloudflare": "^4.8.0",
+		"@unocss/preset-rem-to-px": "^0.65.1",
+		"@unocss/reset": "^0.65.1",
 		"clsx": "^2.1.1",
-		"discord-api-types": "0.37.101",
-		"dotenv": "^16.4.5",
+		"discord-api-types": "0.37.111",
+		"dotenv": "^16.4.7",
 		"js-cookie": "^3.0.5",
 		"svelte-fa": "^4.0.3",
-		"tailwind-merge": "^2.5.3"
+		"tailwind-merge": "^2.5.5"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^3.2.5",
-		"@sveltejs/kit": "^2.6.1",
-		"@sveltejs/vite-plugin-svelte": "^4.0.0-next.7",
+		"@sveltejs/adapter-auto": "^3.3.1",
+		"@sveltejs/kit": "^2.10.1",
+		"@sveltejs/vite-plugin-svelte": "^5.0.2",
 		"@types/eslint": "^9.6.1",
 		"@types/js-cookie": "^3.0.6",
-		"@types/node": "^22.7.4",
-		"@typescript-eslint/eslint-plugin": "^8.8.0",
-		"@typescript-eslint/parser": "^8.8.0",
-		"@unocss/svelte-scoped": "^0.63.3",
-		"eslint": "^8.57.1",
+		"@types/node": "^22.10.2",
+		"@typescript-eslint/eslint-plugin": "^8.18.0",
+		"@typescript-eslint/parser": "^8.18.0",
+		"@unocss/svelte-scoped": "^0.65.1",
+		"eslint": "^9.16.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.44.1",
-		"prettier": "^3.3.3",
-		"prettier-plugin-svelte": "^3.2.7",
-		"svelte": "5.0.0-next.262",
-		"typescript": "^5.6.2",
-		"unocss": "^0.63.3",
+		"eslint-plugin-svelte": "^2.46.1",
+		"prettier": "^3.4.2",
+		"prettier-plugin-svelte": "^3.3.2",
+		"svelte": "5.11.2",
+		"typescript": "^5.7.2",
+		"unocss": "^0.65.1",
 		"unocss-preset-animations": "^1.1.0",
 		"unocss-preset-shadcn": "^0.3.1",
-		"vite": "^5.4.8"
+		"vite": "^6.0.3"
 	},
 	"packageManager": "pnpm@9.10.0",
 	"engines": {
